fix: add UI improvements and feedbacks
