+++ src/components/cards/WrapperCard.tsx
@@ -1,0 +1,20 @@
+import React from "react";
+import { Card } from "react-daisyui";
+
+type WrapperCardProps = {
+  id: string;
+  children: React.ReactNode;
+};
+
+const WrapperCard = ({ id, children }: WrapperCardProps) => {
+  return (
+    <Card
+      id={id}
+      bordered={false}
+      className="flex h-fit flex-col rounded-[14px] bg-white shadow-md">
+      <Card.Body className="gap-4 p-4 md:gap-6 md:p-8">{children}</Card.Body>
+    </Card>
+  );
+};
+
+export default WrapperCard;
