+++ styles/common.ts
@@ -1,0 +1,2 @@
+export const sidePadding = "px-4";
+export const rootInput = "flex flex-col gap-4";
